% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_collections}
\alias{get_collections}
\title{Get Hilltop Collections}
\usage{
get_collections(endpoint = "http://envdata.tasman.govt.nz/data.hts?")
}
\arguments{
\item{endpoint}{A url for the Hilltop endpoint.}
}
\value{
A tibble of collections.
}
\description{
Function to get Collections from Hilltop Server.
}
\examples{
get_collections("http://envdata.tasman.govt.nz/data.hts?")
}
