% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data_collection}
\alias{get_data_collection}
\title{Get Hilltop Data for a Collection}
\usage{
get_data_collection(
  endpoint = "http://envdata.tasman.govt.nz/data.hts?",
  collection,
  method = NA,
  interval = "1 day",
  from = NA,
  to = NA,
  time_interval = NA,
  alignment = "00:00"
)
}
\arguments{
\item{endpoint}{A url for the Hilltop endpoint.}

\item{collection}{A string for the collection.}

\item{method}{A string for the statistic computed, currently only tested for Total, Average and Extrema.}

\item{interval}{A string for the interval over which the "method" is calculated, e.g. "1 day".}

\item{from}{A string for the start date.}

\item{to}{A string for the end date.}

\item{time_interval}{A string for the relative time, e.g. P1Y (Past 1 year). See Hilltop Server documentation}

\item{alignment}{See Hilltop Server documentation.}
}
\value{
A tibble containing the data for each site comprising the collection.
}
\description{
Function to get Data for a Collection from Hilltop Server.
}
\examples{
get_data_collection("http://envdata.tasman.govt.nz/data.hts?", collection = "Rainfall",
      method = "Total", interval = "1 hour", from = "20220101", to = "202207031")
get_data_collection("http://envdata.tasman.govt.nz/data.hts?", collection = "ActiveFlowSites",
      method = "Extrema", interval = "15 minutes", from = "20220101", to = "202207031")
}
