% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilltopserver_functions.R
\name{get_data_site_measurement}
\alias{get_data_site_measurement}
\title{Get Data for a Site and Measurement}
\usage{
get_data_site_measurement(
  endpoint = "http://envdata.tasman.govt.nz/data.hts?",
  site,
  measurement,
  method = NA,
  interval = "1 day",
  from = NA,
  to = NA,
  time_interval = NA,
  alignment = "00:00"
)
}
\arguments{
\item{endpoint}{A url for the Hilltop endpoint.}

\item{site}{A string for the site.}

\item{measurement}{A string for the measurement.}

\item{method}{A string for the statistic computed, currently only tested for Total, Average and Extrema.}

\item{interval}{A string for the interval over which the "method" is calculated, e.g. "1 day".}

\item{from}{A string for the start date.}

\item{to}{A string for the end date.}

\item{time_interval}{A string for the relative time, e.g. P1Y (Past 1 year). See Hilltop Server documentation}

\item{alignment}{See Hilltop Server documentation.}
}
\value{
A tibble containing the data for each site comprising the collection.
}
\description{
Function to get Data for a Site and Measurement from Hilltop Server.
}
\examples{
get_data_site_measurement("http://envdata.tasman.govt.nz/data.hts?",
  site = "AQ Richmond Central at Plunket", measurement = "PM2.5 (24 Hour)",
  interval = "1 hour", from = "20220101", to = "202207031"
)
get_data_site_measurement("http://envdata.tasman.govt.nz/data.hts?",
  site = "HY Richmond Weather at TDC Roof", measurement = "Rainfall",
  method = "Total", interval = "15 minutes", from = "20220101", to = "20220731"
)
}
