% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilltopserver_functions.R
\name{get_gaugings}
\alias{get_gaugings}
\title{Get Gaugings for a Site}
\usage{
get_gaugings(endpoint = "http://envdata.tasman.govt.nz/data.hts?", site)
}
\arguments{
\item{endpoint}{A url for the Hilltop endpoint.}

\item{site}{A string for the site.}
}
\value{
A tibble containing the gaugings for the site.
}
\description{
Function to get Gaugings for a Site from Hilltop Server.
}
\examples{
get_gaugings("http://envdata.tasman.govt.nz/data.hts?",
  site = "HY Aorere at Devils Boots"
)
}
