% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilltopserver_functions.R
\name{get_ratings}
\alias{get_ratings}
\title{Get Ratings for a Site}
\usage{
get_ratings(endpoint = "http://envdata.tasman.govt.nz/data.hts?", site)
}
\arguments{
\item{endpoint}{A url for the Hilltop endpoint.}

\item{site}{A string for the site.}
}
\value{
A tibble containing the ratings for the site.
}
\description{
Function to get Ratings for a Site from Hilltop Server.
}
\examples{
get_ratings("http://envdata.tasman.govt.nz/data.hts?",
  site = "HY Aorere at Devils Boots"
)
}
