% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sites}
\alias{get_sites}
\title{Get Hilltop Sites}
\usage{
get_sites(endpoint = "http://envdata.tasman.govt.nz/data.hts?", latlong = TRUE)
}
\arguments{
\item{endpoint}{A url for the Hilltop endpoint.}

\item{latlong}{A logical, TRUE returns EPSG:4326 (WGS) coordinates , EPSG:2193 (NZTM) otherwise.}
}
\value{
A tibble of sites and their corresponding coordinates.
}
\description{
Function to get Sites from Hilltop Server.
}
\examples{
get_sites("http://envdata.tasman.govt.nz/data.hts?")
get_sites("http://envdata.tasman.govt.nz/data.hts?", latlong = FALSE)
}
