% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilltopserver_functions.R
\name{get_version}
\alias{get_version}
\title{Get Hilltop Server Version}
\usage{
get_version(endpoint = "http://envdata.tasman.govt.nz/data.hts?")
}
\arguments{
\item{endpoint}{A url for the Hilltop endpoint.}
}
\value{
Version text.
}
\description{
Function to get Hilltop Server Version.
}
\examples{
get_version("http://envdata.tasman.govt.nz/data.hts?")
}
